fix: remove bypass from the list of applications
Remove `:bypass` from the list of applications in `mix.exs`. Having it
in the list of applications causes issues in other project's test suites
where an error would be raised, because the bypass application could not
be found.
The test helper module is also updated to make sure that the bypass
application is started when running the test suite.